pyproject.toml
[tool.poetry]
name = "whisper-audio-live-transcription"
version = "1.0.0"
description = ""
authors = ["Zarif Aziz <zarif@aziz.com>"]
readme = "README.md"
include = []

[tool.poetry.dependencies]
python = "^3.10"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
whispercpp = {git = "https://github.com/stlukey/whispercpp.py"}
pyaudio = "^0.2.14"
pydub = "^0.25.1"
wavio = "^0.0.8"
sounddevice = "^0.4.6"
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
